Ability to check symbols in linux

+++ SymbolsLoader/SymbolsManager.h
@@ -1,8 +1,9 @@
 #ifndef SYMBOLSMANAGER_H
 #define SYMBOLSMANAGER_H
+#include <cstddef>
 
 namespace sm {
-    bool getFunctionInfo(const void* address, char* functionInfo, const unsigned int maxlen);
+    bool getFunctionInfo(const void* address, char* functionInfo, const size_t maxlen);
 }
